Separate methods into two new classes for both testing and cohesion purposes
